requirements.txt
Beaker>=1.6.3
Mako>=0.7.1
pymongo>=2.2.1
ply>=3.4
tornado>=2.4
ipaddr>=2.1.10
celery>3.0
python-memcached>=1.48
redis>=2.7.0
hiredis>=0.1.1
